requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
requests
python-dotenv
